+++ src/gazeArbiter/include/iCub/trackerThread.h
@@ -33,7 +33,7 @@
 #include <yarp/os/Thread.h>
 #include <yarp/os/Time.h>
 #include <yarp/sig/Image.h>
-
+#include <yarp/os/Semaphore.h>
 #include <cv.h>
 #include <highgui.h>
 
@@ -168,12 +168,12 @@ public:
                 ImageOf<PixelMono> &tmp = imgMonoPrev;    // image containing the template
 
                 // specify the searching area
-                search_roi.x=(std::max)(0,(std::min)(img.width()-search_roi.width,point.x-(search_roi.width>>1)));
-                search_roi.y=(std::max)(0,(std::min)(img.height()-search_roi.height,point.y-(search_roi.height>>1)));
+                search_roi.x=(std::max)(0,(std::min)((int) (img.width()-search_roi.width), point.x-(search_roi.width>>1)));
+                search_roi.y=(std::max)(0,(std::min)((int) (img.height()-search_roi.height), point.y-(search_roi.height>>1)));
 
                 // specify the template area
-                template_roi.x=(std::max)(0,(std::min)(tmp.width()-template_roi.width,point.x-(template_roi.width>>1)));
-                template_roi.y=(std::max)(0,(std::min)(tmp.height()-template_roi.height,point.y-(template_roi.height>>1)));
+                template_roi.x=(std::max)(0,(std::min)((int) (tmp.width()-template_roi.width), point.x-(template_roi.width>>1)));
+                template_roi.y=(std::max)(0,(std::min)((int) (tmp.height()-template_roi.height),point.y-(template_roi.height>>1)));
 
                 // perform tracking with template matching
                 float ftmp;
